Otway-Rees: added SecurityProps module and proved secrecy of conversation key
The secrecy property shows that the conversation key is secret as long as all
involved parties, i.e., initiator, responder and auth server, are not corrupted.
However, this property should ultimately not hold true, as there are several known
interception attacks on the protocol that reveal the final conversation key.
Therefore, it is necessary to model the respective attacks first before having
a deeper look into the security properties.