Otway-Rees: implemented msg 2 interception attack (execution results in error)
The attack cannot work due to the usage of unique 'ad' values in aead enc/dec
to identify the encrypted parts and due to the structure of messages that
contain a concrete type of encval, which is different for msg2 and msg3.
To make the attack work, adaptions on the message impl. and use of 'ad' values
is required.