Otway-Rees: changed message impl. to make msg2 interception attack work
The message impl. had to be adapted such that the tag of the encrypted message
parts is not part of the value that will be encrypted. This was needed such
that initiator and responder interpret the encrypted values from the second
message as the ones encrypted by the server in the third message, containing
the shared conversation key between initiator and responder.