Otway-Rees: implemented impersonate responder attack
Implemented the attack in which the attacker makes use of the fact that the server
does not verify whether the combination of conv. id and principal names in the
encrypted parts match the unencrypted part. Thus, the attacker can replace the
responder name in the unencrypted part and gets a key intended for the initiator
and him, while the initiator will believe it is for him and the actual responder
(that is not involved in the protocol run at all).