Otway-Rees: renamed some attacker Lemmas with misleading names
Since the Otway-Rees protocol cannot achieve authentication, the Lemma names
were misleading in that they tried to proof that one party believes of the
authentication of the other or that both parties believe in mutual auth.
The wording "believe to talk to" is more appropriate in this context, as it
is decided based on the terms stored in the respective final principal states.