Otway-Rees: added new directory with fixed Otway-Rees protocol
The changes to the protocol are as minimalistic as possible and the resulting
protocol does still not achieve mutual authentication. As this was not explicitly
mentioned as a goal of the original protocol (according to the protocol description),
it only remains to show that the keys stored in the initiator and respectively,
in the responder state, are secret.