DS: trigger send key event in third protocol step
The sign pred had to be adjusted such that F* can infer that the certify event
was triggered by the server with the actual pub keys of the initiator and responder.
The event pred has been refined for the send key event, such that for all honest
or dishonest principals, the clock count will always be <= second message receive
delay and if the server is honest, a preceeding certify event has been triggered.