DS: adapted event/sign preds, added pke preds and adapted third step accordingly
The changes were necessary to get to the point that we can show key secrecy for
the communication key on the responder side. However, the fourth step is not
provable yet because I am not able to show that the event pred holds for the
certify event (within the fourth step).