DS: added SecurityProps module with lemmas for key secrecy and replay mitigation
The replay mitigation property could be shown in the model, based on the timestamp
and clock counts from the accept key and certify events of the responder and,
respectively, the server. Key secrecy on the initiator side was trivial, but
not so for the responder side (there is still a predicate in the post-cond.
that should be eliminated, if possible). For initiator authentication there is
still a quantifier in the predicate of the accept key event that needs to be
eliminated as well.