DS: adapted DS.Clock module such that attacker can request new clock instance
The declaration of the clock_new function has been moved to DS.Clock.fsti such
that it can be called without having to friend the DS.Clock module.
However, a clock can now only be created with the current timestamp instead of
an arbitrary timestamp, limiting the possibilities to pass a wrong clock to
a protocol function for the attacker. On the other hand it enables the attacker
to issue false certificates after corrupting the server (which he could do in
reality).