YLM: tried to proof initiator authentication
I tried to use the knowledge of the responder about his nonce n_b to show
that there is a 'fwd key' event of the initiator matching his 'recv key' event.
However, there is a problem in the fourth protocol step where F* does not know
whether the responder session is valid anymore (maybe because of too much
nesting ??). I solved it by assuming the required property about n_b, which is
the exact label of n_b and should definitely be known to B. I still need to
figure out why F* can't infer that and remove the assume statement later.